fixed test suite not exiting. and ensured smooth working of all tests
