+++ backend/app.js
@@ -1,0 +1,26 @@
+// app.js
+import express from 'express';
+import dotenv from 'dotenv';
+import cors from 'cors';
+import postRoutes from './routes/posts.js';
+import userRoutes from './routes/users.js';
+
+dotenv.config();
+
+const app = express();
+
+// Middleware
+app.use(express.json());
+app.use(cors());
+
+// Logging middleware
+app.use((req, res, next) => {
+  console.log(req.path, req.method);
+  next();
+});
+
+// Routes
+app.use('/api/posts', postRoutes);
+app.use('/api/user', userRoutes);
+
+export default app;
