+++ backend/server.js
@@ -4,22 +4,10 @@ import mongoose from 'mongoose';
 import cors from 'cors';
 import postRoutes from './routes/posts.js';
 import userRoutes from './routes/users.js';
+import app from './app.js';
 
 dotenv.config();
-const port = process.env.PORT;
-
-const app = express();
-app.use(express.json());
-app.use(cors());
-
-app.use((req, res, next) => {
-  console.log(req.path, req.method);
-  next();
-});
-
-app.use('/api/posts', postRoutes);
-app.use('/api/user', userRoutes);
-
+const port = process.env.PORT || 3000;
 const connectDB = async () => {
   try {
     mongoose.set('strictQuery', true);
@@ -42,5 +30,3 @@ connectDB()
     });
   })
   .catch((err) => console.log(err));
-
-export { app };
